+++ lib/bots/commands_server/util/current_app_utils.rb
