+++ lib/config/persisted_hash.rb
